made changes in main branch without realizing

+++ src/commands/channelonly.ts
@@ -1,53 +1,53 @@
-import ICommandArguments from "../interfaces/ICommandArguments";
-import channelCommandSchema from "../models/channel-commands";
+import { ICallbackObject, ICommand } from '../..'
+import channelCommandSchema from '../models/channel-commands'
 
 export = {
   minArgs: 1,
   expectedArgs: '<Command name> [Channel tags OR "none"]',
-  cooldown: "2s",
-  requiredPermissions: ["ADMINISTRATOR"],
+  cooldown: '2s',
+  requiredPermissions: ['ADMINISTRATOR'],
   guildOnly: true,
-  description: "Makes a command only work in some channels.",
-  category: "Configuration",
-  callback: async (options: ICommandArguments) => {
-    const { message, args, instance } = options;
-    const { guild } = message;
+  description: 'Makes a command only work in some channels.',
+  category: 'Configuration',
+  callback: async (options: ICallbackObject) => {
+    const { message, args, instance } = options
+    const { guild } = message
 
-    const { messageHandler } = instance;
+    const { messageHandler } = instance
 
-    let commandName = (args.shift() || "").toLowerCase();
-    const command = instance.commandHandler.getICommand(commandName);
+    let commandName = (args.shift() || '').toLowerCase()
+    const command = instance.commandHandler.getICommand(commandName)
 
-    if (!command) {
+    if (!command || !command.names) {
       message.reply(
-        messageHandler.get(guild, "UNKNOWN_COMMAND", {
+        messageHandler.get(guild, 'UNKNOWN_COMMAND', {
           COMMAND: commandName,
         })
-      );
-      return;
+      )
+      return
     }
 
-    commandName = command.names[0];
-    const action = args[0];
+    commandName = command.names[0]
+    const action = args[0]
 
-    if (action && action.toLowerCase() === "none") {
+    if (action && action.toLowerCase() === 'none') {
       const results = await channelCommandSchema.deleteMany({
         guildId: guild?.id,
         command: commandName,
-      });
+      })
 
       if (results.n === 0) {
-        message.reply(messageHandler.get(guild, "NOT_CHANNEL_COMMAND"));
+        message.reply(messageHandler.get(guild, 'NOT_CHANNEL_COMMAND'))
       } else {
-        message.reply(messageHandler.get(guild, "NO_LONGER_CHANNEL_COMMAND"));
+        message.reply(messageHandler.get(guild, 'NO_LONGER_CHANNEL_COMMAND'))
       }
 
-      return;
+      return
     }
 
     if (message.mentions.channels.size === 0) {
-      message.reply(messageHandler.get(guild, "NO_TAGGED_CHANNELS"));
-      return;
+      message.reply(messageHandler.get(guild, 'NO_TAGGED_CHANNELS'))
+      return
     }
 
     await channelCommandSchema.findOneAndUpdate(
@@ -65,13 +65,13 @@ export = {
       {
         upsert: true,
       }
-    );
+    )
 
     message.reply(
-      messageHandler.get(guild, "NOW_CHANNEL_COMMAND", {
+      messageHandler.get(guild, 'NOW_CHANNEL_COMMAND', {
         COMMAND: commandName,
-        CHANNELS: args.join(" "),
+        CHANNELS: args.join(' '),
       })
-    );
+    )
   },
-};
+} as ICommand
